v0.2 Confirmation Messages and addition of JS

+++ src/Subscriptions/views.py
@@ -7,7 +7,6 @@ from .bgprocess import calculate_monthly_total, calculate_weekly_total, calculat
 def dashboard(request):
     if request.user.is_authenticated:
         user_subscriptions = Subscription.objects.filter(user=request.user).order_by('provider__name')
-        print(calculate_yearly_total(user_subscriptions))
         context = {
             "subscriptions": user_subscriptions,
             "weekly_total": calculate_weekly_total(user_subscriptions),
@@ -16,31 +15,39 @@ def dashboard(request):
         }
         return render(request, 'subscriptions/dashboard.html', context)
     else:
+        messages.error(request, 'You must be logged in to view this page', extra_tags='danger')
         return redirect('users_login')
 
 def create_subscription(request):
-    if request.method == 'POST':
-        form = CreateSubscription(request.POST)
-        if form.is_valid():
-            form.save(commit=False)
-            form.instance.user = request.user
-            form.save()
-            messages.success(request, 'Subscription added successfully', extra_tags='success')
-            return redirect('subs_dashboard')
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            form = CreateSubscription(request.POST)
+            if form.is_valid():
+                form.save(commit=False)
+                form.instance.user = request.user
+                form.save()
+                messages.success(request, 'Subscription added successfully', extra_tags='success')
+                return redirect('subs_dashboard')
+            else:
+                messages.error(request, 'Subscription could not be added', extra_tags='danger')
+                return render(request, 'subscriptions/create.html', {'form': form})
         else:
-            messages.error(request, 'Subscription could not be added', extra_tags='danger')
+            form = CreateSubscription()
             return render(request, 'subscriptions/create.html', {'form': form})
     else:
-        form = CreateSubscription()
-        return render(request, 'subscriptions/create.html', {'form': form})
+        messages.error(request, 'You must be logged in to view this page', extra_tags='danger')
+        return redirect('users_login')
 
 def delete_subscription(request, id):
-    subscription = Subscription.objects.get(id=id)
-    if subscription.user == request.user:
-        subscription.delete()
-        messages.success(request, 'Subscription deleted successfully', extra_tags='success')
+    if request.user.is_authenticated:
+        subscription = Subscription.objects.get(id=id)
+        if subscription.user == request.user:
+            subscription.delete()
+            messages.success(request, 'Subscription deleted successfully', extra_tags='success')
+        else:
+            messages.error(request, 'Subscription could not be deleted', extra_tags='danger')
+        return redirect('subs_dashboard')
     else:
-        messages.error(request, 'Subscription could not be deleted', extra_tags='danger')
-    return redirect('subs_dashboard')
-
+        messages.error(request, 'You must be logged in to view this page', extra_tags='danger')
+        return redirect('users_login')
